+++ packages/react/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* Add `hash`, `query`, and `state` props to `<Link>`, remove `details`.
 * Remove `<Link active>`.
 * Update `<Active>` to use a render-invoked `children` prop. Removes the `merge`, `details`, and `extra` props.
 * Use the new React Context API released in React 16.3. This forces the minimum React version for `@curi/react` projects to `16.3`.